requirements.txt
aiofiles==23.2.1
aiogram==3.20.0.post0
aiohttp==3.9.1
aiosignal==1.3.2
aiosqlite==0.21.0
alembic==1.13.1
annotated-types==0.7.0
attrs==25.3.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
frozenlist==1.7.0
greenlet==3.2.3
idna==3.10
importlib_metadata==8.7.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
loguru==0.7.3
magic-filter==1.0.12
Mako==1.3.10
MarkupSafe==3.0.2
multidict==6.4.4
notifiers==1.3.6
propcache==0.3.2
pydantic==2.11.5
pydantic-settings==2.9.1
pydantic_core==2.33.2
python-dotenv==1.1.0
referencing==0.36.2
requests==2.32.4
rpds-py==0.25.1
SQLAlchemy==2.0.41
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.4.0
yarl==1.20.1
zipp==3.23.0
